Refactored and fixed time, exceptoins, and zahed examples
