--- examples/java/exceptions/run.bat
@@ -1,11 +1,0 @@
-@echo off
-call ..\env.bat
-
-if not exist Exceptions.class (
-  echo  Compiling Exceptions.java
-  javac Exceptions.java
-)
-
-java Exceptions
-
-pause
