--- examples/java/exceptions2/run.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-. ../env.sh
-
-run Exceptions2
-
