--- examples/java/time/run.bat
@@ -1,11 +1,0 @@
-@echo off
-call ..\env.bat
-
-if not exist Time.class (
-  echo  Compiling Time.java
-  javac Time.java
-)
-
-java Time
-
-pause
