--- examples/java/time/run.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-. ../env.sh
-
-run Time
-
