+++ src/examples/java/README.md
@@ -26,9 +26,12 @@ In one line:
 find . -name "*.java" -print | xargs javac -cp /usr/lib/swi-prolog/lib/jpl.jar
 ```
 
-Tu run any example, we need to provide the location of `jpl.jar`.
+Tu **run the examples** we need to:
+ 
+ * provide the location of `jpl.jar`, always. 
+ * run it within its folder so it can find any corresponding resource `.pl` (Prolog source) files, if any.
 
-For example, to run a particular example `X` that does not require any resources (e.g., consulting a Prolog file) we can simply do:
+For example, to run a particular example `X` that does _not_ require any resources (e.g., consulting a Prolog file) we can simply do:
 
 ```bash
 java -cp .:/usr/lib/swi-prolog/lib/jpl.jar  org.jpl7.examples.X
@@ -44,15 +47,24 @@ java -cp ../../../../:/usr/lib/swi-prolog/lib/jpl.jar org.jpl7.examples.family.F
 
 ## Examples
 
-Exceptions
-	demonstrates how an uncaught Prolog exception is turned
-	into a Java exception (also uncaught, in this example)
+## Exceptions
 
-Exceptions2
-	demonstrates how an uncaught Prolog exception is turned
-	into a Java exception, which in this example is caught,
-	converted to a String and printed on System.out (and
-	also shown in a Swing treeview gadget)
+Demonstrates how an uncaught Prolog exception is turned into a Java exception, which in this example is caught in Java.
+
+```
+[ssardina@Thinkpad-X1 java]$ java -cp .:/usr/lib/swi-prolog/lib/jpl.jar  org.jpl7.examples.Exceptions
+calling
+
+?- X is Y.
+
+in Prolog to force a Prolog 'instantiation_error' exception,
+which should be returned via Java as an uncaught org.jpl7.PrologException in thread "main".
+
+BINGO! The following exception was thrown: 
+	PrologException: error(instantiation_error, context(':'(system, '/'(is, 2)), _2))
+```
+
+
 
 ### Family
 
@@ -99,28 +111,67 @@ factorial(10) = 3628800
 ```
 
 
-Time
-	runs a set of timing experiments, designed to gauge the speed
-	with which terms are exchanged between Java and Prolog
+### Time
+	
+Runs a set of timing experiments, designed to gauge the speed with which terms are exchanged between Java and Prolog.
+
+```bash
+[ssardina@Thinkpad-X1 time]$ java -cp ../../../../:/usr/lib/swi-prolog/lib/jpl.jar org.jpl7.examples.time.Time
+test 0...passed
+treesdone
+traversedone
+noopdone
+noop_nobinddone
+noop_binddone
+
+num_trials: 10
+tree_depth: 10 (= 1025 terms)
+
+test_0		test_1		test_2		test_3		test_4		
+0		5		2		2		2		
+0		0		1		1		2		
+0		0		1		1		3		
+0		0		1		1		2		
+1		0		1		1		2		
+0		0		0		1		1		
+0		0		1		1		2		
+0		0		1		1		2		
+0		0		2		1		2		
+0		0		1		1		3		
+
+test_0: avg: 0ms		0ms/term
+test_1: avg: 0ms		0ms/term
+test_2: avg: 1ms		0.001ms/term
+test_3: avg: 1ms		0.001ms/term
+test_4: avg: 2ms		0.002ms/term
+```
 
 ### Versions
 
 The `org.jpl7.Versions` class example tries to discover which versions (e.g. `7.6.0-alpha`) of each JPL library (i.e. Prolog, C and Java) are installed, and congratulates you if they are the same. It also reports the home and version of the SWIPL engine being used.
 
-Zahed
-	creates a trivial-in-Prolog-but-horrible-in-JPL1.x query goal
-	and calls it against a trivial Prolog database; this would be much
-	nicer if it used recent facilities for creating queries from
-	Prolog source text fragments
+### Zahed
+	
+Creates a trivial-in-Prolog-but-horrible-in-JPL query goal and calls it against a trivial Prolog database (file `zahed.pl`).
 
-NB each folder contains a Java application which should be run within
-its folder so it can find any corresponding .pl (Prolog source) files:
-Windows users can use 'compile.bat' to (re)compile each demo, and
-'run.bat' to run it.
+The query term is built "from first principles", by composing the various data types one by one. Of course, in the new JPL, one uses recent facilities for creating query terms from Prolog source text fragments (namely, `Term.textToTerm()`).
+
+```
+[ssardina@Thinkpad-X1 zahed]$ java -cp ../../../../:/usr/lib/swi-prolog/lib/jpl.jar org.jpl7.examples.zahed.Zahed
+starting...
+The query constructed is: gen([t(c, q, []), t(v, [], a)], A)
+ok
+finished
+```
 
 ----
+
+# Contributions
 
 Paul Singleton (paul.singleton@bcs.org.uk)
 July 2003
 January 2004 (revised)
 December 2004 (not revised much apart from this line)
+
+Sebastian Sardina (ssardina@gmail.com)
+June 2020
