+++ src/examples/java/org/jpl7/examples/zahed/run.sh
