+++ src/examples/java/sources.txt
@@ -1,5 +1,8 @@
 ./org/jpl7/examples/family/FamilyMT.java
 ./org/jpl7/examples/family/Family.java
+./org/jpl7/examples/zahed/Zahed.java
 ./org/jpl7/examples/factorial/Factorial.java
+./org/jpl7/examples/Exceptions.java
+./org/jpl7/examples/time/Time.java
 ./org/jpl7/examples/test/Test.java
 ./org/jpl7/examples/Versions.java
